Add action reserve term
