+++ src/Http/Controllers/ConsultationAPIController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace EscolaLms\Consultations\Http\Controllers;
+
+use EscolaLms\Consultations\Http\Controllers\Swagger\ConsultationAPISwagger;
+use EscolaLms\Core\Http\Controllers\EscolaLmsBaseController;
+
+class ConsultationAPIController extends EscolaLmsBaseController implements ConsultationAPISwagger
+{
+    public function reportTerm(int $id)
+    {
+
+    }
+}
