+++ src/Http/Controllers/Swagger/ConsultationAPISwagger.php
@@ -1,424 +1,6 @@
 <?php
 namespace EscolaLms\Consultations\Http\Controllers\Swagger;
 
-use EscolaLms\Consultations\Http\Requests\ListConsultationsRequest;
-use EscolaLms\Consultations\Http\Requests\StoreConsultationRequest;
-use EscolaLms\Consultations\Http\Requests\UpdateConsultationRequest;
-use Illuminate\Http\JsonResponse;
-
 interface ConsultationAPISwagger
 {
-
-    /**
-     * @OA\Get(
-     *      path="/api/consultations",
-     *      summary="Get a listing of the Consultations.",
-     *      tags={"Consultations"},
-     *      description="Get all Consultations",
-     *      @OA\Parameter(
-     *          name="order_by",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *              enum={"finished_at", "started_at", "created_at", "name", "base_price"}
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="order",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *              enum={"ASC", "DESC"}
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="page",
-     *          description="Pagination Page Number",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="number",
-     *               default=1,
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="per_page",
-     *          description="Pagination Per Page",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="number",
-     *               default=15,
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="name",
-     *          description="Consultation name %LIKE%",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="base_price",
-     *          description="Consultation base_price >= ",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="integer",
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="status",
-     *          description="Consultation status == ",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *          ),
-     *      ),
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json"
-     *          ),
-     *          @OA\Schema(
-     *              type="object",
-     *              @OA\Property(
-     *                  property="success",
-     *                  type="boolean"
-     *              ),
-     *              @OA\Property(
-     *                  property="data",
-     *                  type="array",
-     *                  @OA\Items(ref="#/components/schemas/Consultation")
-     *              ),
-     *              @OA\Property(
-     *                  property="message",
-     *                  type="string"
-     *              )
-     *          )
-     *      )
-     * )
-     */
-
-    /**
-     * @OA\Get(
-     *      path="/api/admin/consultations",
-     *      @OA\Parameter(
-     *          name="order_by",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *              enum={"finished_at", "started_at", "created_at", "name", "base_price"}
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="order",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *              enum={"ASC", "DESC"}
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="page",
-     *          description="Pagination Page Number",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="number",
-     *               default=1,
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="per_page",
-     *          description="Pagination Per Page",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="number",
-     *               default=15,
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="name",
-     *          description="Consultation name %LIKE%",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="base_price",
-     *          description="Consultation base_price >= ",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="integer",
-     *          ),
-     *      ),
-     *      @OA\Parameter(
-     *          name="status",
-     *          description="Consultation status == ",
-     *          required=false,
-     *          in="query",
-     *          @OA\Schema(
-     *              type="string",
-     *          ),
-     *      ),
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json"
-     *          ),
-     *          @OA\Schema(
-     *              type="object",
-     *              @OA\Property(
-     *                  property="success",
-     *                  type="boolean"
-     *              ),
-     *              @OA\Property(
-     *                  property="data",
-     *                  type="array",
-     *                  @OA\Items(ref="#/components/schemas/Consultation")
-     *              ),
-     *              @OA\Property(
-     *                  property="message",
-     *                  type="string"
-     *              )
-     *          )
-     *      )
-     * )
-     */
-    public function index(ListConsultationsRequest $listConsultationsRequest): JsonResponse;
-
-    /**
-     * @OA\Post(
-     *      path="/api/admin/consultations",
-     *      summary="Store a newly created Consultation in storage",
-     *      tags={"Admin Consultations"},
-     *      description="Store Consultation",
-     *      security={
-     *          {"passport": {}},
-     *      },
-     *      @OA\RequestBody(
-     *          required=true,
-     *          @OA\MediaType(
-     *              mediaType="application/json",
-     *              @OA\Schema(ref="#/components/schemas/Consultation")
-     *          ),
-     *      ),
-     *      @OA\Response(
-     *          response=201,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json"
-     *          ),
-     *          @OA\Schema(
-     *              type="object",
-     *              @OA\Property(
-     *                  property="success",
-     *                  type="boolean"
-     *              ),
-     *              @OA\Property(
-     *                  property="data",
-     *                  ref="#/components/schemas/Consultation"
-     *              ),
-     *              @OA\Property(
-     *                  property="message",
-     *                  type="string"
-     *              )
-     *          )
-     *      )
-     * )
-     */
-    public function store(StoreConsultationRequest $storeConsultationRequest): JsonResponse;
-
-    /**
-     * @OA\Get(
-     *      path="/api/admin/consultations/{id}",
-     *      summary="Display the specified Consultation",
-     *      tags={"Admin Consultations"},
-     *      description="Get Consultation",
-     *      security={
-     *         {"passport": {}},
-     *      },
-     *      @OA\Parameter(
-     *          name="id",
-     *          description="id of Consultation",
-     *          @OA\Schema(
-     *             type="integer",
-     *         ),
-     *          required=true,
-     *          in="path"
-     *      ),
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json"
-     *          ),
-     *          @OA\Schema(
-     *              type="object",
-     *              @OA\Property(
-     *                  property="success",
-     *                  type="boolean"
-     *              ),
-     *              @OA\Property(
-     *                  property="data",
-     *                  ref="#/components/schemas/Consultation"
-     *              ),
-     *              @OA\Property(
-     *                  property="message",
-     *                  type="string"
-     *              )
-     *          )
-     *      )
-     * )
-     */
-
-    /**
-     * @OA\Get(
-     *      path="/api/consultations/{id}",
-     *      summary="Display the specified Consultation",
-     *      tags={"Consultations"},
-     *      description="Get Consultation",
-     *      @OA\Parameter(
-     *          name="id",
-     *          description="id of Consultation",
-     *          @OA\Schema(
-     *             type="integer",
-     *         ),
-     *          required=true,
-     *          in="path"
-     *      ),
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json"
-     *          ),
-     *          @OA\Schema(
-     *              type="object",
-     *              @OA\Property(
-     *                  property="success",
-     *                  type="boolean"
-     *              ),
-     *              @OA\Property(
-     *                  property="data",
-     *                  ref="#/components/schemas/Consultation"
-     *              ),
-     *              @OA\Property(
-     *                  property="message",
-     *                  type="string"
-     *              )
-     *          )
-     *      )
-     * )
-     */
-    public function show(int $id): JsonResponse;
-
-    /**
-     * @OA\Put(
-     *      path="/api/admin/consultations/{id}",
-     *      summary="Update the specified Consultation in storage",
-     *      tags={"Admin Consultations"},
-     *      description="Update Consultation",
-     *      security={
-     *          {"passport": {}},
-     *      },
-     *      @OA\Parameter(
-     *          name="id",
-     *          description="id of Consultation",
-     *          @OA\Schema(
-     *             type="integer",
-     *         ),
-     *          required=true,
-     *          in="path"
-     *      ),
-     *      @OA\RequestBody(
-     *          required=true,
-     *          @OA\MediaType(
-     *              mediaType="application/json",
-     *              @OA\Schema(ref="#/components/schemas/Consultation")
-     *          )
-     *      ),
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json"
-     *          ),
-     *          @OA\Schema(
-     *              type="object",
-     *              @OA\Property(
-     *                  property="success",
-     *                  type="boolean"
-     *              ),
-     *              @OA\Property(
-     *                  property="data",
-     *                  ref="#/components/schemas/Consultation"
-     *              ),
-     *              @OA\Property(
-     *                  property="message",
-     *                  type="string"
-     *              )
-     *          )
-     *      )
-     * )
-     */
-    public function update(int $id, UpdateConsultationRequest $updateConsultationRequest): JsonResponse;
-
-    /**
-     * @OA\Delete(
-     *      path="/api/admin/consultations/{id}",
-     *      summary="Remove the specified Consultation from storage",
-     *      tags={"Admin Consultations"},
-     *      description="Delete Consultation",
-     *      security={
-     *          {"passport": {}},
-     *      },
-     *      @OA\Parameter(
-     *          name="id",
-     *          description="id of Consultation",
-     *          @OA\Schema(
-     *             type="integer",
-     *         ),
-     *          required=true,
-     *          in="path"
-     *      ),
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json"
-     *          ),
-     *          @OA\Schema(
-     *              type="object",
-     *              @OA\Property(
-     *                  property="success",
-     *                  type="boolean"
-     *              ),
-     *              @OA\Property(
-     *                  property="data",
-     *                  type="string"
-     *              ),
-     *              @OA\Property(
-     *                  property="message",
-     *                  type="string"
-     *              )
-     *          )
-     *      )
-     * )
-     */
-    public function destroy(int $id): JsonResponse;
 }
